refactor project story hooks
